moved ParkPreview to its own file. added iamge to preivew. changed content target to prevent merge conflicts
